Move auth to web module
