--- app/controllers/auth/base_controller.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module Auth
-  # Base controller for authentication (signup, session, oauths). Place common methods here.
-  class BaseController < ApplicationController; end
-end
