+++ app/controllers/web/auth/base_controller.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# Base controller for authentication (signup, session, oauths). Place common methods here.
+module Web
+  module Auth
+    class BaseController < Web::BaseController; end
+  end
+end
