--- app/views/auth/user_sessions/new.html.haml
@@ -1,4 +1,0 @@
-.container
-  %h2
-    = t('.login')
-  = render 'form'
