--- app/views/auth/users/edit.html.haml
@@ -1,9 +1,0 @@
-%h2
-  = t('.edit_user_label')
-.col-md-6
-  = render 'form'
-
-.col-md-6
-  = link_to t('.delete_user'), user_path(current_user),
-                          method: :delete, class: 'btn btn-danger',
-                          data: { confirm: t('.delete_user_confirmation') }
