--- app/views/auth/users/new.html.haml
@@ -1,4 +1,0 @@
-%h2
-  = t('.registration_label')
-
-= render 'form'
