+++ app/views/web/auth/user_sessions/new.html.haml
@@ -1,0 +1,4 @@
+.container
+  %h2
+    = t('auth.user_sessions.new.login')
+  = render 'form'
