+++ app/views/web/auth/users/edit.html.haml
@@ -1,0 +1,9 @@
+%h2
+  = t('auth.users.edit.edit_user_label')
+.col-md-6
+  = render 'form'
+
+.col-md-6
+  = link_to t('auth.users.edit.delete_user'), user_path(current_user),
+                          method: :delete, class: 'btn btn-danger',
+                          data: { confirm: t('auth.users.edit.delete_user_confirmation') }
