+++ app/views/web/auth/users/new.html.haml
@@ -1,0 +1,4 @@
+%h2
+  = t('auth.users.new.registration_label')
+
+= render 'form'
